refactor: event-detail feature を prototype.jsx から分離 (#8)
- features/event-detail/EventDetailPanel.jsx: EventDetailPanel を移設
  (Meet/Zoom リンク、添付資料、Markdown description を維持)
- features/event-detail/EventDetailPanel.test.jsx: 9件の smoke test
  - タイトル/時刻/プロジェクトバッジ、Meet/Zoom/リンクなしの出し分け
  - attachments リスト、description 空時のプレースホルダー
  - backdrop クリックで onClose
- prototype.jsx から EventDetailPanel 宣言を削除 (約246行削減)

これで task-detail (#7) と event-detail が両方 features 配下に揃い、
将来 shared/ui/BottomSheet に共通化する基盤が整う。
